docs: Turn the http bench into an example
